pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-mcp-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0",
    "starlette",
    "uvicorn[standard]",
    "google-api-python-client",
    "google-auth-oauthlib",
    "pymupdf",
    "python-docx",
    "chromadb",
    "sentence-transformers",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "pytest-asyncio",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src --cov=config --cov-report=term-missing"

[tool.coverage.run]
omit = [
    "tests/*",
    "src/mcp_server/server.py",  # __main__ block excluded via pragma
]
